requirements.txt
PyGithub
PyYAML
requests
tqdm 